Publish jazzer plugin
This commit adds publishing of the Jazzer plugin. It will NOT be
published on the Gradle plugin portal, but on Maven Central only.
This shouldn't be an issue, except that users won't be able to
find it when searching on the plugin portal (similarly to the
native build tools).

It may be worth checking before a 1.0.0 release with a beta version
or something, since the setup is a bit special.

Fixes #100