Rework build logic
This commit reworks how the model class is "shared" between the annotation
processor and the jazzer plugin. Instead of using a symlink, the code is
now generated by a common build plugin. In practice, the 2 modules use the
same record, but in a different package. It's also cleaner in the sense
that there's no dependency between the plugin and the annotation processor,
which would have led to too many classes on classpath.

Fixes #101